+++ internal/rabbitmq/provider_test.go
