+++ internal/rabbitmq/resource_permissions.go
