+++ internal/rabbitmq/resource_queue.go
