+++ internal/rabbitmq/resource_user.go
