+++ internal/rabbitmq/resource_user_test.go
@@ -5,8 +5,8 @@ import (
 	"regexp"
 	"testing"
 
-	"github.com/rfd59/terraform-provider-rabbitmq/acceptance"
-	"github.com/rfd59/terraform-provider-rabbitmq/acceptance/check"
+	"github.com/rfd59/terraform-provider-rabbitmq/internal/acceptance"
+	"github.com/rfd59/terraform-provider-rabbitmq/internal/acceptance/check"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 )
