+++ internal/rabbitmq/util.go
