+++ internal/rabbitmq/util_test.go
